Use octocrab for GraphQL queries (#187)
Remove the manual use of reqwest::Client for graphql
queries and use octocrab instead.

This removes the need to specify the github api url
multiple times. We instead leverage the global octocrab
instance for which we configure the api url once.

This is beneficial for a cleaner implementation of
Github Enterprise support (#158), as we can configure
the base url once.